Add fast path in sphere collision function
Since `t_min` is always positive, it is impossible for a ray to hit a
sphere if it is facing away from it. This can be checked by using the
dot product of the ray direction and the direction between the ray and
the sphere, which was already being calculated.